Add StateMachine loading and saving
Each state is able to save it's own (unique) name to a variable. This variable can be saved and later be passed to a (fershly created) Builder (implementing the IStateLoader interface) to ask him to search for a state with that (unique) name within it's created states. If the builder does not find a state claiming the name it will continue asking each of it's IStateMachineFactories (which now extends IStateLoader) if they know a state claiming $name. The factories will ask their Builders and so on (in a recursive manner) until either a state claiming the name was found or no other States / IStateMachineBuilders are known. In the latter case null is returned, else the found state is returned.