Add drawUnit optimization to only draw units within dirtyRect
After increasing the number of units the performance dropped
significanty and allthough clipping was acgtive drawUnits took 25ms
to complete. Adding a bounds check to drawUnits on each iteration
which checks if the unit's place lies within the dirtyRect and only drawing 
it if it does reduced drawing time to roughly 2.5ms.